requirements.txt
# Core ML & LLM
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.25.0
sentence-transformers>=2.2.0

# PDF Processing
PyPDF2>=3.0.0

# Vector Database
chromadb>=0.4.20

# Web UI
gradio>=4.0.0

# API (optional, but included)
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Config & Utilities
PyYAML>=6.0